Refactor: Rename Head component to Header and update import paths
